+++ lib/browser/browser.rb
@@ -31,6 +31,7 @@ require_relative "yandex"
 require_relative "sputnik"
 require_relative "snapchat"
 require_relative "duck_duck_go"
+require_relative "samsung_browser"
 
 require_relative "bot"
 require_relative "bot/empty_user_agent_matcher"
@@ -72,6 +73,7 @@ module Browser
       Yandex,               # must be placed before Chrome and Safari
       Sputnik,              # must be placed before Chrome and Safari
       DuckDuckGo,           # must be placed before Chrome and Safari
+      SamsungBrowser,       # must be placed before Chrome and Safari
       Chrome,
       Safari,
       MicroMessenger,