+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Add Samsung browser.
+- Add The Knowledge AI bot to the bot list.
+- Add Go HttpClient to the bot list.
 
 ## 4.0.0