+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Add Samsung browser.
 - Add Google Image Proxy to the bot list.
+- Add The Knowledge AI bot to the bot list.
+- Add Go HttpClient to the bot list.
+- Fix MicroMessenger detection on Android
 
 ## 4.0.0